Reverse order of declaration of new relationships to enhance layout
I noticed that the previous layout was not optimal in terms
of edges crossing which could be avoided by moving nodes
in different positions. It seemed that the layout was strongly
influenced by the order of the entities in the declaration of
the relationship, which is confirmed after the change: reversing
the order of entities had a great impact on the layout.