Update schema in SQL to create the database
This is the latest schema, retrieved using `show tables`
and `show create table table_name` for each table.

These declarations combine fields and constraints, which
are not longer declared separately using `ALTER TABLE`.

Based on `show create table ...` output, I added the
`IF NOT EXISTS` clause and removed the configuration of
initial counter value for `AUTO_INCREMENT`, since the
intent of this script is to recreate an empty database.