Set AUTO_INCREMENT to MAX(id)+1 before inserts with id=0
I left the value '0' in id of new inserts, to get the next value
from the AUTO_INCREMENT sequence.

However, when inserting into a new database, the AUTO_INCREMENT
is initially set to 1 and updated after each new record is inserted
to one more than the highest value.

As a result, inserts done before the highest current id received
a value too low, possibly used in existing records, which failed
to be inserted due to the duplicate key value.