Add initial export for France participations, from ipcc-acquire
The data was exported from `giec` database using the script
`answer.sh` in the project `ipcc-acquire`, temporarily modified
to connect to the previous database `giec` instead of `ipcc`.

The file `answer12/list-french-author-participations.tsv`
generated by `answer.sh` was converted from TSV to CSV with
the script `scripts/tsv2csv.awk` part of this project.

From this point, I will copy the query which generates this data
to the export scripts, to keep this data updated with each export.